Clear code #2
