+++ tableTest/Base.lproj/Main.storyboard
@@ -2,7 +2,7 @@
 <document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="17156" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="BYZ-38-t0r">
     <device id="retina5_5" orientation="portrait" appearance="light"/>
     <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="17125"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="17126"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="System colors in document resources" minToolsVersion="11.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
@@ -16,24 +16,32 @@
                         <rect key="frame" x="0.0" y="0.0" width="414" height="736"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Wol-ql-BMs">
-                                <rect key="frame" x="192" y="103" width="30" height="30"/>
-                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Wol-ql-BMs">
+                                <rect key="frame" x="192" y="130" width="30" height="30"/>
                                 <state key="normal" title="Get"/>
                                 <connections>
                                     <action selector="getButtonAction:" destination="BYZ-38-t0r" eventType="touchUpInside" id="Wnc-M6-B43"/>
                                 </connections>
                             </button>
-                            <textField opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="left" contentVerticalAlignment="center" text="Send text" borderStyle="roundedRect" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="IOG-H8-Chg">
+                            <textField opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" text="Send text" borderStyle="roundedRect" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="IOG-H8-Chg">
                                 <rect key="frame" x="42" y="61" width="330" height="34"/>
-                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                <constraints>
+                                    <constraint firstAttribute="width" constant="330" id="6OO-nX-no2"/>
+                                    <constraint firstAttribute="height" constant="34" id="fJC-Gm-Tof"/>
+                                </constraints>
                                 <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                 <textInputTraits key="textInputTraits"/>
+                                <connections>
+                                    <action selector="getTextOutAction:" destination="BYZ-38-t0r" eventType="editingDidEnd" id="hw4-Tu-tkG"/>
+                                </connections>
                             </textField>
-                            <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" fixedFrame="YES" text="ReceiveText: " textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="WSW-5H-xH9">
-                                <rect key="frame" x="42" y="178" width="330" height="153"/>
-                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                            <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" text="ReceiveText: " textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="WSW-5H-xH9">
+                                <rect key="frame" x="42" y="205" width="330" height="153"/>
                                 <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                                <constraints>
+                                    <constraint firstAttribute="width" constant="330" id="9Sl-yS-WB4"/>
+                                    <constraint firstAttribute="height" constant="153" id="gK3-rH-bQC"/>
+                                </constraints>
                                 <color key="textColor" systemColor="labelColor"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                 <textInputTraits key="textInputTraits" autocapitalizationType="sentences"/>
@@ -41,6 +49,17 @@
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                        <constraints>
+                            <constraint firstItem="IOG-H8-Chg" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" constant="61" id="3u8-o3-GKb"/>
+                            <constraint firstItem="Wol-ql-BMs" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" constant="192" id="6H1-Fe-TFk"/>
+                            <constraint firstItem="WSW-5H-xH9" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" constant="42" id="Bfs-aA-UQy"/>
+                            <constraint firstItem="WSW-5H-xH9" firstAttribute="top" secondItem="Wol-ql-BMs" secondAttribute="bottom" constant="45" id="JMj-HB-np6"/>
+                            <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="WSW-5H-xH9" secondAttribute="trailing" constant="42" id="KmW-Df-pSC"/>
+                            <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="Wol-ql-BMs" secondAttribute="trailing" constant="192" id="ele-T4-v82"/>
+                            <constraint firstItem="IOG-H8-Chg" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" constant="42" id="gxI-Bv-d1X"/>
+                            <constraint firstItem="Wol-ql-BMs" firstAttribute="top" secondItem="IOG-H8-Chg" secondAttribute="bottom" constant="35" id="kKc-bs-6cU"/>
+                            <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="IOG-H8-Chg" secondAttribute="trailing" constant="42" id="wEt-y0-Guk"/>
+                        </constraints>
                     </view>
                     <connections>
                         <outlet property="textIn" destination="WSW-5H-xH9" id="vcn-8z-SVT"/>
