+++ tableTest/Enums/URLs.swift
@@ -1,0 +1,11 @@
+//
+//  URLs.swift
+//  tableTest
+//
+//  Created by Andrew Iskandarov on 28.10.2020.
+//
+
+import Foundation
+enum URLs {
+  static let  baseURL = "https://jsonplaceholder.typicode.com/todos/"
+}
